- finished the connection with the doc and the F2 feature    (all thats left is to fix the docbooks) - fixed some HEAVY bugs (weird that i didnt spot them before) - switched some methods from int to double (new policy)
svn path=/trunk/kdeedu/kturtle/; revision=360870

+++ src/executer.h
@@ -54,11 +54,11 @@ class Executer : public QObject
 		void MessageDialog(QString text);
 	
 		void Clear();
-		void Go(int x, int y);
-		void GoX(int x);
-		void GoY(int y);
-		void Forward(int x);
-		void Backward(int x);
+		void Go(double x, double y);
+		void GoX(double x);
+		void GoY(double y);
+		void Forward(double x);
+		void Backward(double x);
 		void Direction(double x);
 		void TurnLeft(double x);
 		void TurnRight(double x);
